onboarding-lambda/requirements.txt
opentelemetry-api>=1.26.0
opentelemetry-sdk>=1.26.0
opentelemetry-exporter-otlp-proto-grpc>=1.26.0
grpcio>=1.62.0
protobuf>=4.25.0